requirements.txt
absl-py==1.4.0
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.3
asgiref==3.5.2
astunparse==1.6.3
async-timeout==4.0.2
attrdict==2.0.1
attrs==22.2.0
autopep8==2.0.2
Babel==2.9.1
beautifulsoup4==4.12.1
boto3==1.26.107
botocore==1.29.107
bottle==0.12.25
cachetools==5.3.0
captcha==0.4
certifi==2022.5.18.1
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
clr-loader==0.2.5
colorama==0.4.6
contourpy==1.0.7
coreapi==2.3.3
coremltools==6.3.0
coreschema==0.0.4
cryptography==40.0.1
cycler==0.11.0
defusedxml==0.7.1
Django==4.0.5
django-ckeditor==6.5.1
django-cleanup==7.0.0
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-extensions==3.2.1
django-filter==23.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-modeladmin-reorder==0.3.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
emoji==2.2.0
filelock==3.11.0
flatbuffers==23.3.3
fonttools==4.39.3
frozenlist==1.3.3
future==0.18.3
gast==0.4.0
google==3.0.0
google-api-core==1.34.0
google-auth==2.17.2
google-auth-oauthlib==1.0.0
google-cloud-vision==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.0
grpcio==1.53.0
grpcio-status==1.48.2
h5py==3.8.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
jax==0.4.8
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
keras==2.12.0
kiwisolver==1.4.4
kraken==2.0.8
libclang==16.0.0
lxml==4.9.2
magic-filter==1.0.9
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
ml-dtypes==0.1.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==2.0.0
pefile==2023.2.7
Pillow==9.4.0
proto-plus==1.22.2
protobuf==3.20.3
proxy-tools==0.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pyinstaller==5.9.0
pyinstaller-hooks-contrib==2023.1
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.3
pyTelegramBotAPI==4.10.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==1.0.0
python3-openid==3.2.0
pythonnet==3.0.1
pytz==2022.7.1
pywebview==4.0.2
pywin32-ctypes==0.2.0
regex==2023.3.23
requests==2.27.1
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.1
rsa==4.9
ruamel.yaml==0.17.21
s3transfer==0.6.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn==0.0.post1
social-auth-app-django==5.2.0
social-auth-core==4.4.1
soupsieve==2.4
sqlparse==0.4.2
sympy==1.11.1
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
threadpoolctl==3.1.0
tkinterweb==3.17.7
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.2.3
wrapt==1.14.1
yarl==1.8.2